+++ neurokit2/eog/__init__.py
@@ -7,15 +7,16 @@ from .eog_eventrelated import eog_eventrelated
 from .eog_features import eog_features
 from .eog_findpeaks import eog_findpeaks
 from .eog_intervalrelated import eog_intervalrelated
+from .eog_peaks import eog_peaks
 from .eog_plot import eog_plot
 from .eog_process import eog_process
-
 
 __all__ = [
     "eog_rate",
     "eog_clean",
     "eog_features",
     "eog_findpeaks",
+    "eog_peaks",
     "eog_process",
     "eog_plot",
     "eog_eventrelated",