+++ neurokit2/rsp/rsp_fixpeaks.py
@@ -12,17 +12,17 @@ def rsp_fixpeaks(peaks, troughs=None):
     Parameters
     ----------
     peaks : list or array or DataFrame or Series or dict
-        The samples at which the inhalation peaks occur. If a dict or a DataFrame is passed, it is
-        assumed that these containers were obtained with `rsp_findpeaks()`.
+        The samples at which respiration peaks (exhalation onsets) occur. If a dict or a DataFrame is passed,
+        it is assumed that these containers were obtained with `rsp_findpeaks()`.
     troughs : list or array or DataFrame or Series or dict
-        The samples at which the inhalation troughs occur. If a dict or a DataFrame is passed, it is
-        assumed that these containers were obtained with `rsp_findpeaks()`.
+        The samples at which respiration troughs (inhalation onsets) occur. If a dict or a DataFrame is passed,
+        it is assumed that these containers were obtained with `rsp_findpeaks()`.
 
     Returns
     -------
     info : dict
-        A dictionary containing additional information, in this case the samples at which inhalation
-        peaks and exhalation troughs occur, accessible with the keys "RSP_Peaks", and "RSP_Troughs", respectively.
+        A dictionary containing additional information, in this case the samples at which inhalation onsets
+        and exhalation onsets occur, accessible with the keys "RSP_Troughs" and "RSP_Peaks", respectively.
 
     See Also
     --------