+++ neurokit2/rsp/rsp_plot.py
@@ -62,9 +62,9 @@ def rsp_plot(rsp_signals, sampling_rate=None):
     ax0.plot(x_axis, rsp_signals["RSP_Raw"], color="#B0BEC5", label="Raw", zorder=1)
     ax0.plot(x_axis, rsp_signals["RSP_Clean"], color="#2196F3", label="Cleaned", zorder=2, linewidth=1.5)
 
-    ax0.scatter(x_axis[peaks], rsp_signals["RSP_Clean"][peaks], color="red", label="Inhalation Peaks", zorder=3)
+    ax0.scatter(x_axis[peaks], rsp_signals["RSP_Clean"][peaks], color="red", label="Exhalation Onsets", zorder=3)
     ax0.scatter(
-        x_axis[troughs], rsp_signals["RSP_Clean"][troughs], color="orange", label="Exhalation Troughs", zorder=4
+        x_axis[troughs], rsp_signals["RSP_Clean"][troughs], color="orange", label="Inhalation Onsets", zorder=4
     )
 
     # Shade region to mark inspiration and expiration.